+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ------------------
 - prerender for SEO
 - added `bitcoin:` anchor on checkout (so you can click and your client will open up)
+- created settings wrapper around `localStorage`
 
 0.0.1 / 2014-10-09
 ------------------